+++ packages/dds/tree/src/core/tree/cursor.ts
@@ -484,78 +484,3 @@ export function inCursorNode<T, TCursor extends ITreeCursor = ITreeCursor>(
 	cursor.exitNode();
 	return result;
 }
-export type NodeComparator = (cursor: ITreeCursorSynchronous) => boolean;
-
-type FieldComparator = (cursor: ITreeCursorSynchronous) => boolean;
-
-function buildFieldComparator(
-	expectedLength: number,
-	nodeComparators: NodeComparator[],
-): FieldComparator {
-	return (cursor: ITreeCursorSynchronous): boolean => {
-		if (cursor.getFieldLength() !== expectedLength) {
-			return false;
-		}
-		for (let index = 0; index < expectedLength; index++) {
-			cursor.enterNode(index);
-			const comparator = nodeComparators[index];
-			if (comparator?.(cursor) !== true) {
-				cursor.exitNode();
-				return false;
-			}
-			cursor.exitNode();
-		}
-		return true;
-	};
-}
-/**
- * Walks the cursor at its current node position and builds a comparator that can test whether another
- * cursor position has the same structure and values.
- *
- * @remarks
- * Fields are compared by key (not iteration order), so this is safe regardless of cursor field ordering.
- * Missing fields in the compared node will cause a mismatch.
- *
- * The cursor must be in Nodes mode. After this call. the cursor is restored to its original position.
- *
- */
-export function buildNodeComparator(cursor: ITreeCursorSynchronous): NodeComparator {
-	const expectedValue: Value = cursor.value;
-	const fieldComparators: { key: FieldKey; compare: FieldComparator }[] = [];
-
-	for (let inField = cursor.firstField(); inField; inField = cursor.nextField()) {
-		const key = cursor.getFieldKey();
-		const length = cursor.getFieldLength();
-		const nodeComparators: NodeComparator[] = [];
-
-		for (let index = 0; index < length; index++) {
-			cursor.enterNode(index);
-			nodeComparators.push(buildNodeComparator(cursor));
-			cursor.exitNode();
-		}
-
-		fieldComparators.push({
-			key,
-			compare: buildFieldComparator(length, nodeComparators),
-		});
-	}
-	// Note: if firstField() returned false, we're already back in Nodes mode.
-	// If it returned true and we iterated to exhaustion, nextField returned
-	// false, which also put us back in Nodes mode.
-
-	return (other: ITreeCursorSynchronous): boolean => {
-		if (other.value !== expectedValue) {
-			return false;
-		}
-		for (const { key, compare } of fieldComparators) {
-			other.enterField(key);
-			if (compare(other) !== true) {
-				other.exitField();
-				return false;
-			}
-			other.exitField();
-		}
-
-		return true;
-	};
-}
