+++ packages/dds/tree/src/core/tree/index.ts
@@ -29,8 +29,6 @@ export {
 	inCursorNode,
 	CursorMarker,
 	isCursor,
-	type NodeComparator,
-	buildNodeComparator,
 } from "./cursor.js";
 export type {
 	ProtoNodes,
